requirements.txt
# 相控阵雷达仿真平台 - Python依赖
# 版本: 2.0

# ============================================================================
# 核心依赖
# ============================================================================

# 数值计算
numpy>=1.24.0
scipy>=1.10.0
numba>=0.57.0

# 数据处理
pandas>=2.0.0

# ============================================================================
# Web框架
# ============================================================================

fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
aiofiles>=23.0
python-multipart>=0.0.5

# ============================================================================
# 数据序列化
# ============================================================================

msgpack>=1.0.0
orjson>=3.9.0

# ============================================================================
# 配置管理
# ============================================================================

pydantic>=2.0.0
python-dotenv>=1.0.0
tomli>=2.0.0

# ============================================================================
# 日志系统
# ============================================================================

loguru>=0.7.0

# ============================================================================
# 异步IO
# ============================================================================

asyncio-mqtt>=0.14.0  # 用于MQTT集成(可选)

# ============================================================================
# 机器学习
# ============================================================================

scikit-learn>=1.2.0

# ============================================================================
# 算法和数据结构
# ============================================================================

heapq>=0.1.0

# ============================================================================
# 测试框架
# ============================================================================

pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0
pytest-mock>=3.11.0

# ============================================================================
# 代码质量工具
# ============================================================================

black>=23.0.0
isort>=5.12.0
pylint>=2.17.0
flake8>=6.0.0
mypy>=1.4.0

# ============================================================================
# 性能分析
# ============================================================================

py-spy>=0.3.14
memory-profiler>=0.61.0

# ============================================================================
# 类型存根
# ============================================================================

types-requests>=2.31.0
types-aiofiles>=23.0.0

# ============================================================================
# 可选依赖（高性能计算）
# ============================================================================

# pyFFTW用于FFT加速
# pyfftw>=0.13.0

# CuPy用于GPU加速（需要CUDA）
# cupy-cuda11x>=12.0.0
